refactor: move away from app directory
