+++ src/components-client/RegisterPage/LoginForm.tsx
@@ -3,9 +3,9 @@
 import Button from "@components-client/Form/Button";
 import Input from "@components-client/Form/Input";
 import { FieldValues, SubmitHandler, useForm } from "react-hook-form";
-import { supabase } from "../../../lib/supabase";
+import { supabase } from "../../lib/supabase";
 
-const RegisterForm = () => {
+const LoginForm = () => {
   // use react-hook-form to manage form state and submission
   const {
     register,
@@ -18,26 +18,14 @@ const RegisterForm = () => {
     console.log("executed onSubmit!");
     // use Supabase auth here :D
     try {
-      const takenPlayerName = await supabase
-        .from("PlayerTable")
-        .select("*")
-        .eq("playerName", formData.playerName);
-      if (takenPlayerName?.data?.length)
-        throw new Error("A player already exists with this name.");
-
-      const takenEmail = await supabase
-        .from("PlayerTable")
-        .select("*")
-        .eq("playerEmail", formData.playerEmail);
-      if (takenEmail?.data?.length)
-        throw new Error("A player already exists with this email.");
-
       if (formData.playerEmail) {
         const { error } = await supabase.auth.signInWithOtp({
           email: formData.playerEmail,
-          options: { data: { playerName: formData.playerName } },
+          options: {
+            shouldCreateUser: false,
+          },
         });
-        if (error) throw new Error("For some reason this didn't work :shrug:");
+        if (error) throw new Error(error.message);
       }
       alert("Check your email for the login link!");
     } catch (error) {
@@ -55,12 +43,6 @@ const RegisterForm = () => {
   return (
     <>
       <form className="flex flex-col gap-4" onSubmit={handleSubmit(onSubmit)}>
-        <Input
-          title="Player Name"
-          placeholder="First Last"
-          type="text"
-          {...register("playerName", { required: true })}
-        />
         <Input
           title="Player Email (use vincit.fi)"
           placeholder="first.last@vincit.fi"
@@ -88,11 +70,11 @@ const RegisterForm = () => {
           className="button-primary"
           isLoading={isSubmitting}
         >
-          Sign Up!
+          Login!
         </Button>
       </form>
     </>
   );
 };
 
-export default RegisterForm;
+export default LoginForm;
