+++ src/components-client/RegisterPage/RegisterForm.tsx
@@ -3,9 +3,9 @@
 import Button from "@components-client/Form/Button";
 import Input from "@components-client/Form/Input";
 import { FieldValues, SubmitHandler, useForm } from "react-hook-form";
-import { supabase } from "../../../lib/supabase";
+import { supabase } from "../../lib/supabase";
 
-const LoginForm = () => {
+const RegisterForm = () => {
   // use react-hook-form to manage form state and submission
   const {
     register,
@@ -18,14 +18,26 @@ const LoginForm = () => {
     console.log("executed onSubmit!");
     // use Supabase auth here :D
     try {
+      const takenPlayerName = await supabase
+        .from("PlayerTable")
+        .select("*")
+        .eq("playerName", formData.playerName);
+      if (takenPlayerName?.data?.length)
+        throw new Error("A player already exists with this name.");
+
+      const takenEmail = await supabase
+        .from("PlayerTable")
+        .select("*")
+        .eq("playerEmail", formData.playerEmail);
+      if (takenEmail?.data?.length)
+        throw new Error("A player already exists with this email.");
+
       if (formData.playerEmail) {
         const { error } = await supabase.auth.signInWithOtp({
           email: formData.playerEmail,
-          options: {
-            shouldCreateUser: false,
-          },
+          options: { data: { playerName: formData.playerName } },
         });
-        if (error) throw new Error(error.message);
+        if (error) throw new Error("For some reason this didn't work :shrug:");
       }
       alert("Check your email for the login link!");
     } catch (error) {
@@ -43,6 +55,12 @@ const LoginForm = () => {
   return (
     <>
       <form className="flex flex-col gap-4" onSubmit={handleSubmit(onSubmit)}>
+        <Input
+          title="Player Name"
+          placeholder="First Last"
+          type="text"
+          {...register("playerName", { required: true })}
+        />
         <Input
           title="Player Email (use vincit.fi)"
           placeholder="first.last@vincit.fi"
@@ -70,11 +88,11 @@ const LoginForm = () => {
           className="button-primary"
           isLoading={isSubmitting}
         >
-          Login!
+          Sign Up!
         </Button>
       </form>
     </>
   );
 };
 
-export default LoginForm;
+export default RegisterForm;
