+++ src/components-client/RegisterPage/index.tsx
@@ -1,8 +1,9 @@
 import RegisterForm from "./RegisterForm";
 import LoginForm from "./LoginForm";
-const Register = () => {
+import PageLayout from "@components-client/PageLayout";
+const RegisterPage = () => {
   return (
-    <>
+    <PageLayout segment="Register">
       <div className="my-8">
         <h3 className="mb-4 text-xl font-bold">Login</h3>
         <LoginForm />
@@ -12,7 +13,7 @@ const Register = () => {
         <h3 className="mb-4 text-xl font-bold">Register</h3>
         <RegisterForm />
       </div>
-    </>
+    </PageLayout>
   );
 };
-export default Register;
+export default RegisterPage;
