+++ src/pages/register/index.tsx
@@ -1,0 +1,10 @@
+import RegisterPage from "@components-client/RegisterPage";
+
+// eslint-disable-next-line import/no-anonymous-default-export, react/display-name
+export default () => {
+  return (
+    <>
+      <RegisterPage />
+    </>
+  );
+};
